feat: Native attributes for status indicator
